Add Eigen build option with Intel MKL
