+++ environment.devenv.yml
@@ -9,6 +9,8 @@ dependencies:
   - cmake>=3.13, <3.18
   - ninja
   - eigen
+  - mkl
+  - mkl-include
   - boost==1.74
   - gxx_linux-64  # [linux]
   - ccache  # [unix]
